Updated UI for order history, unified API service classes

+++ app/src/main/java/com/eggbucket/eggbucket_b2c/User.kt
@@ -10,7 +10,7 @@ data class User(
     val gender: String,
     val age: String,
     val email: String,
-    val totalExpenditure: Int,
+    val totalExpenditure: Double,
     val totalOrders: Int,
     val name: String,
     @SerializedName("addresses") val userAddresses: List<UserAddress>
